+++ lib/ansible/ruby/modules/custom/cloud/core/amazon/ec2_ami.rb
@@ -10,6 +10,9 @@ module Ansible
     module Modules
       class Ec2_ami
         include Helpers::Aws
+
+        remove_existing_validations :state
+        validates :state, inclusion: { :in => [:present, :absent], :message => "%{value} needs to be :present, :absent" }, allow_nil: true
       end
     end
   end