+++ lib/ansible/ruby/modules/custom/network/basics/get_url.rb
@@ -1,11 +1,13 @@
 # frozen_string_literal: true
 # VALIDATED_CHECKSUM: v5T3dbovbGBE0ljyTcaYnW+gG0ChWYC8CHAat1jZPP8=
 require 'ansible/ruby/modules/generated/core/network/basics/get_url'
+require 'ansible/ruby/modules/helpers/file_attributes'
 
 module Ansible
   module Ruby
     module Modules
       class Get_url
+        include Helpers::FileAttributes
         remove_existing_validations :headers
         # @return [Hash, nil] Add custom HTTP headers to a request in the format "key:value,key:value"
         validates :headers, type: Hash