+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * `lineinfile` module - added missing `validate` parameter
 * `meta` should now work properly like `command` as a free form module
 * `with_items` supports splat style array parsing
+* `vars` now supported in task and blocks
 
 # 1.0.6 - 14 October 2016
 * Allow sharing variables between tasks (#29)